Force tests in a module to be run by the same worker (#2177)
* Force tests in a module to be run by the same worker


* Put link to pytest-xdist docs.


* Remove flyteidl setup make target


* Set flyteidl lower bound to 1.10.7b4


* Set a default value for PYTEST_OPTS


* Fix matching of error in test_python_entrypoint


* Disable pytest-xdist temporarily


* Remove default from makefile temporarily


* Move test that touches the local cache to test_local_cache


* Mock environment using mock.patch.dict in test_serialization


* Re-enable pytest-xdist


* Lint


* Move patching of os.environ closer to where it's required


* Check that a different key is not usable in test_secrets_manager_default


---------
