+++ tests/flytekit/unit/configuration/test_internal.py
@@ -48,7 +48,6 @@ def test_client_secret_location():
     assert platform_cfg.auth_mode == AuthType.CLIENTSECRET.value
 
 
-@mock.patch.dict("os.environ")
 def test_client_secret_env_var():
     cfg = get_config_file(os.path.join(os.path.dirname(os.path.realpath(__file__)), "configs/sample.yaml"))
     secret_env_var = Credentials.CLIENT_CREDENTIALS_SECRET_ENV_VAR.read(cfg)
@@ -60,10 +59,10 @@ def test_client_secret_env_var():
     secret_env_var = Credentials.CLIENT_CREDENTIALS_SECRET_ENV_VAR.read(cfg)
     assert secret_env_var == "FAKE_SECRET_NAME"
 
-    os.environ["FAKE_SECRET_NAME"] = "fake_secret_value"
-    platform_cfg = PlatformConfig.auto(cfg)
-    assert platform_cfg.client_credentials_secret == "fake_secret_value"
-    assert platform_cfg.auth_mode == AuthType.CLIENTSECRET.value
+    with mock.patch.dict(os.environ, {"FAKE_SECRET_NAME": "fake_secret_value"}):
+        platform_cfg = PlatformConfig.auto(cfg)
+        assert platform_cfg.client_credentials_secret == "fake_secret_value"
+        assert platform_cfg.auth_mode == AuthType.CLIENTSECRET.value
 
 
 def test_read_file_if_exists():
